requirements/base.txt
beautifulsoup4==4.4.1
Django==1.8.7
Markdown==2.6.4
Pillow==3.1.0
cmsplugin-news==0.4.4
django-allauth==0.24.1
django-braces==1.8.1
django-classy-tags==0.7.0
django-cms==3.2.0
django-crispy-forms==1.5.2
django-compressor==2.0
django-gravatar2==1.4.0
django-timezones==0.2
django-configurations==1.0
dj-database-url==0.2.2
django-statici18n==1.1.5
django-sortedm2m==1.1.1
# Strict due to API compatibility issues
django-mptt==0.8.0

django-markup-deprecated==0.0.3
django-reversion==1.10.0
django-sekizai==0.9.0
django-taggit==0.17.6

djangocms-admin-style==1.0.7
djangocms-googlemap==0.4.0
djangocms-text-ckeditor==2.8.1
djangocms-stacks==0.2.3
djangocms-style==1.6.0
djangocms-link==1.7.1
djangocms-inherit==0.1.1
djangocms-snippet==1.7.1

django-userprofiles==0.4
easy-thumbnails==2.3
html5lib==0.9999999
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
gunicorn==19.4.5
psycopg2==2.6.1

lxml==3.5.0


# tablib for export support
tablib==0.10.0

# Search integration
django-haystack==2.4.1

-e git+https://github.com/kliment/pymill.git@f55cab94826168765a5a7d60ffed2641fe89b27d#egg=pymill

# For now we also include Whoosh here. In the future this will
# be moved to dev.txt
Whoosh==2.7.0

# Redis as caching backend
django-redis==4.3.0
hiredis==0.2.0
redis==2.10.5

# Celery for async tasks (pdf generation)
celery==3.1.19

# Documentation
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
